+++ include/link_as_c.hpp
@@ -1,0 +1,9 @@
+// 
+// Copyright 2022 Clemens Cords
+// Created on 07.05.22 by clem (mail@clemens-cords.com)
+//
+
+#pragma once
+
+#define BEGIN_C_LINKAGE extern "C" {
+#define END_C_LINKAGE }
