+++ telescope.hpp
@@ -1,0 +1,15 @@
+#pragme once
+
+#include <telescope.h>
+#include <include/common.hpp>
+#include <include/bullet_interface.hpp>
+#include <include/physics_object.hpp>
+#include <include/texture.hpp>
+#include <include/vertex.hpp>
+#include <include/vulkan_interface.hpp>
+
+
+
+
+
+
